Make submodule for estimation and refactor
